refact: функции generate_payload delete_courier_data перенесены в модуль helpers
