+++ conftest.py
@@ -1,31 +1,13 @@
 import pytest
 import random
-import string
+
+import helpers
 
 from faker import Faker
 from datetime import datetime, timedelta
 
 from endpoints.courier_endpoints import CourierEndpoints
 from endpoints.order_endpoints import OrderEndpoints
-
-
-def generate_payload():
-    def generate_random_string(length):
-        letters = string.ascii_lowercase
-        random_string = ''.join(random.choice(letters) for _ in range(length))
-        return random_string
-
-    login = generate_random_string(10)
-    password = generate_random_string(10)
-    first_name = generate_random_string(10)
-
-    payload = {
-        "login": login,
-        "password": password,
-        "firstName": first_name
-    }
-
-    return payload
 
 
 @pytest.fixture()
@@ -38,32 +20,26 @@ def order_endpoints():
     return OrderEndpoints()
 
 
-def delete_courier_data(payload):
-    courier_endpoint = CourierEndpoints()
-    id_courier = courier_endpoint.login_courier(payload)
-    courier_endpoint.delete_courier(id_courier)
-
-
 @pytest.fixture()
 def payload_to_create_courier():
-    payload = generate_payload()
+    payload = helpers.generate_payload()
     payload_to_delete = {key: value for key, value in payload.items()}
     del payload_to_delete['firstName']
 
     yield payload
 
-    delete_courier_data(payload_to_delete)
+    helpers.delete_courier_data(payload_to_delete)
 
 
 @pytest.fixture()
 def payload_to_login(courier_endpoints):
-    payload = generate_payload()
+    payload = helpers.generate_payload()
     courier_endpoints.create_courier(payload)
     del payload['firstName']
 
     yield payload
 
-    delete_courier_data(payload)
+    helpers.delete_courier_data(payload)
 
 
 @pytest.fixture()
